Move a bunch of files, fix some stuff
sorry...
