+++ api/controllers/CommandReply/add-reply.js
@@ -1,4 +1,4 @@
-const replyTypes = require('../../../api/hooks/sdtdCommands/replyTypes.js');
+const replyTypes = require('../../../worker/processors/sdtdCommands/replyTypes');
 module.exports = {
   friendlyName: 'Add reply',
   description: '',
