+++ worker/processors/sdtdCommands/command.js
