+++ worker/processors/sdtdCommands/commands/claim.js
