+++ worker/processors/sdtdCommands/commands/gimme.js
