+++ worker/processors/sdtdCommands/commands/help.js
@@ -1,4 +1,4 @@
-let SdtdCommand = require('../command.js');
+const SdtdCommand = require('../command.js');
 
 class help extends SdtdCommand {
   constructor(serverId) {
