+++ worker/processors/sdtdCommands/commands/listTele.js
