+++ worker/processors/sdtdCommands/commands/pay.js
