+++ worker/processors/sdtdCommands/commands/removetele.js
