+++ worker/processors/sdtdCommands/commands/seen.js
