+++ worker/processors/sdtdCommands/commands/shop.js
