+++ worker/processors/sdtdCommands/commands/tele.js
