+++ worker/processors/sdtdCommands/commands/telePublic.js
