+++ worker/processors/sdtdCommands/commands/teleprivate.js
