+++ worker/processors/sdtdCommands/commands/vote.js
