+++ worker/processors/sdtdCommands/customCommand.js
