+++ worker/processors/sdtdCommands/parseArgs.js
